+++ libs/co2/data-access/src/lib/co2-emissions/store/create-co2-forecast-interval.ts
